working on edit sku
